refactor: tracking source of enabling notifications
