+++ packages/shared/src/lib/analytics.ts
@@ -10,7 +10,6 @@ export enum Origin {
   TagsSearch = 'tags search',
   RealTime = 'realtime',
   NonRealTime = 'nonrealtime',
-  NewSourceModal = 'new source modal',
 }
 
 export enum LoginTrigger {
